add XServerPinger
To make it easier to catch deadlocks like in NB#198747, ping the X server
periodically if configured with WINDOW_DEBUG, and print a warning if it
doesn't reply timely.  The pinger runs in a different process than the
compositor and they don't interfere in any way, except that when the
compositor exits, so does the pinger.