exposee desktop in checkStacking()
Booting the gadget we falsely tell the desktop it's VisibilityFullyObscured,
thus it doesn't update (but still gets and reacts to pointer events).
This is because an ALWAYS_MAPPED window is mapped after the desktop's.
checkStacking() correctly determines that new window should be below,
but the damage has been done by activateWindow() and isn't undone.
Make checkStacking undo it now.

 * src/mcompositemanager.cpp (MCompositeManagerPrivate::checkStacking):
   Explicitly expose or unexpose the desktop depending on desktop_up.